Eliminated format.bsh.
