+++ import/marc_local.properties
@@ -13,7 +13,6 @@
 # instead of the built-in SolrMarc functionality found in the .jar file.
 # (by default, the scripts have the same behavior as the built-in functions,
 # but the external scripts are easier to customize to your needs).
-#format = script(format.bsh), getFormat, format_map.properties
 #author                = script(author.bsh), getAuthorsFilteredByRelator(100abcd:700abcd,100,firstAuthorRoles)
 #author_variant        = script(author.bsh), getAuthorInitialsFilteredByRelator(100a:700a,100,firstAuthorRoles)
 #author_fuller         = script(author.bsh), getAuthorsFilteredByRelator(100q:700q,100,firstAuthorRoles)
